refactor: Remove deprecated collection service interface
This commit removes the deprecated `collection.service.interface.ts` file, which is no longer used in the codebase. The removal improves code maintainability by eliminating unnecessary files and reducing confusion.
