+++ src/modules/collection/application/interface/collection.base.interface.ts
@@ -1,0 +1,11 @@
+export interface ICollectionValues {
+  name: string;
+  userId?: string;
+  teamId?: string;
+}
+
+export interface IUpdateCollectionValues extends ICollectionValues {
+  id: string;
+}
+
+export const COLLECTION_SERVICE = 'COLLECTION_SERVICE';
